fix(lazadago):修复一系列接口问题

+++ flexicombo/entity/createflexicombogiftskusrequestentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type CreateFlexiComboGiftSkusRequestEntity struct{
-    ProductId	int	`json:"productId"`
-    SkuId	int	`json:"skuId"`
+    ProductId	int64	`json:"productId"`
+    SkuId	int64	`json:"skuId"`
 }
 func (g CreateFlexiComboGiftSkusRequestEntity) String() string {
     return lib.ObjectToString(g)
