+++ flexicombo/entity/createflexicombosampleskusrequestentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type CreateFlexiComboSampleSkusRequestEntity struct{
-    ProductId	int	`json:"productId"`
-    SkuId	int	`json:"skuId"`
+    ProductId	int64	`json:"productId"`
+    SkuId	int64	`json:"skuId"`
 }
 func (g CreateFlexiComboSampleSkusRequestEntity) String() string {
     return lib.ObjectToString(g)
