+++ flexicombo/entity/updateflexicombogiftskusrequestentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type UpdateFlexiComboGiftSkusRequestEntity struct{
-    ProductId	int	`json:"productId"`
-    SkuId	int	`json:"skuId"`
+    ProductId	int64	`json:"productId"`
+    SkuId	int64	`json:"skuId"`
 }
 func (g UpdateFlexiComboGiftSkusRequestEntity) String() string {
     return lib.ObjectToString(g)
