+++ freeshipping/entity/freeshippingselectedproductlistdatalistresponseentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type FreeShippingSelectedProductListDataListResponseEntity struct{
-    ProductId	int	`json:"product_id"`
-    SkuIds	[]int	`json:"sku_ids"`
+    ProductId	int64	`json:"product_id"`
+    SkuIds	[]int64	`json:"sku_ids"`
 }
 func (g FreeShippingSelectedProductListDataListResponseEntity) String() string {
     return lib.ObjectToString(g)
