+++ product/entity/createproductskulistresponseentity.go
@@ -7,7 +7,7 @@ import (
 type CreateProductSkuListResponseEntity struct{
     SellerSku	string	`json:"seller_sku"`
     ShopSku	string	`json:"shop_sku"`
-    SkuId	int	`json:"sku_id"`
+    SkuId	int64	`json:"sku_id"`
 }
 func (g CreateProductSkuListResponseEntity) String() string {
     return lib.ObjectToString(g)
