+++ sellervoucher/entity/sellervoucherselectedproductlistdatalistresponseentity.go
@@ -6,7 +6,7 @@ import (
 
 type SellerVoucherSelectedProductListDataListResponseEntity struct{
     ProductId	int	`json:"product_id"`
-    SkuIds	[]int	`json:"sku_ids"`
+    SkuIds	[]int64	`json:"sku_ids"`
 }
 func (g SellerVoucherSelectedProductListDataListResponseEntity) String() string {
     return lib.ObjectToString(g)
